api-tech-manager/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
pydantic==2.8.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
httpx==0.27.2
email-validator==2.2.0
python-dotenv==1.0.1
supabase==2.5.5
